agentcore-data-governance/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentcore-data-governance"
version = "0.1.0"
description = "Agentic AI Data Governance Operating Model using AWS Bedrock AgentCore"
requires-python = ">=3.12"

dependencies = [
    "bedrock-agentcore[strands-agents]>=1.0.3",
    "strands-agents>=1.13.0",
    "strands-agents-tools>=0.2.16",
    "pydantic>=2.0.0",
    "python-dotenv>=1.2.1",
    # OpenTelemetry for observability (Requirements: 17.1, 17.2)
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "opentelemetry-propagator-aws-xray>=1.0.0",
    "aws-opentelemetry-distro>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "hypothesis>=6.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.hypothesis]
max_examples = 100
deadline = 5000
suppress_health_check = ["too_slow"]

[tool.setuptools.packages.find]
where = ["."]
include = ["agents*", "models*", "repository*", "services*", "tools*"]
